[ stm port ] using the same SNTP implementation files as TI CC3200 SNTP solution uses, just ported to STM32
