--- src/bsp/platform/stm32f4/sntp.h
@@ -1,32 +1,0 @@
-/******************************************************************************
- *                                                                            *
- * sntp.h                                                                     *
- *                                                                            *
- ******************************************************************************/
-
-#ifndef SNTP_H
-#define SNTP_H
-
-#include <stdint.h>
-
-/******************************************************************************
- *                                                                            *
- *  Definitions                                                               *
- *                                                                            *
- ******************************************************************************/
-
-typedef int32_t   posix_time_t;
-typedef uint32_t  ntp_time_t;
-
-
-/******************************************************************************
- *                                                                            *
- * Externs                                                                    *
- *                                                                            *
- ******************************************************************************/
-
-extern void         sntp_init       (void);
-extern ntp_time_t   sntp_time_ntp   (void); // Seconds since 1900/01/01 00:00:00 
-extern posix_time_t sntp_time_posix (void); // Seconds since 1970/01/01 00:00:00 
-
-#endif /* SNTP_H */
